ocr/requirements.txt
pdf2image>=1.17.0
Pillow>=10.0.0
openai>=1.0.0
click>=8.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-docx>=0.8.11
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0